Add packages. Apply more OOP
