+++ src/main/java/duke/DukeExceptions.java
@@ -1,0 +1,7 @@
+package duke;
+
+public class DukeExceptions extends Exception {
+    public DukeExceptions(String s) {
+        super(s);
+    }
+}
